src/a2c/requirements.txt
gym-super-mario-bros
jupyterlab
ipywidgets
opencv-python
matplotlib
seaborn
ipython
nes_py
pytorch